Faster zero_grad
Setting the gradients to none instead of zeroing out is much faster. The behaviour is not exactly the same, but since we are not expecting actual zero gradients in our weights (it would come from unused parameters, which would probably be unintented), it should be equivalent.

Requires fairly recent pytorch.